# Improvements for TLS with I/O threads