@ranshid 2024-11-10
> Looks good to me, but I'd like @madolson to take a look too.
> 
> Impressing performance result, and this is with more lines deleted than added!
> 
> Does the change have any performance impact without I/O threads enabled?

I would be surprised if there was any performance improvement when io-threads are not used. This will actually cause SSL_pending and  ERR_clear_error to be called multiple times in some scenarios where we are reading multiple times on the read handler. BUT I think those cases are rare and AFAIK are mostly limited to cases like the connSyncReadLine.