@uriyage 2024-11-16
> Looks good to me, but I'd like @madolson to take a look too.
> 
> Impressing performance result, and this is with more lines deleted than added!
> 
> Does the change have any performance impact without I/O threads enabled?

As expected, there is no performance difference without I/O threads as the behavior remains the same.


> I would be surprised if there was any performance improvement when io-threads are not used. This will actually cause SSL_pending and ERR_clear_error to be called multiple times in some scenarios where we are reading multiple times on the read handler. BUT I think those cases are rare and AFAIK are mostly limited to cases like the connSyncReadLine.

The only change is indeed for `connSyncReadLine`, but as we use it to call the read syscall one byte at a time, I believe performance is not our concern in this case.